+++ CHANGELOG.md
@@ -8,6 +8,7 @@ To set the file, use `--config FILE` option
 * **BREAKING** Remove `assert_include?` and `refute_include?` in favor of
 Minitest-compatible `assert_includes` and `refute_includes`.
 * **BREAKING** Rename `TLDR::Assertions::MinitestCompatibility` to `TLDR::MinitestCompatibility` and remove `assert_send`, which [nobody uses](https://github.com/minitest/minitest/issues/668)
+* **BREAKING** Replace `no_emoji` YAML option with `emoji` option. Disable emoji output by default. Add `--emoji` flag for enabling it.
 * Add `--[no-]timeout TIMEOUT` flag and `timeout` YAML option. To enable the
 TLDR Classic™ default of 1.8s, specify `--timeout` from the CLI or `timeout: true`
 in YAML. To specify a custom timeout of 42.3 seconds, flag `--timeout 42.3` or